A Maze
This was the third project in Udacity's VR Developer nanodegree. The purpose of this project was to gain experience with C# scripting for Unity and create a game logic. This is the fully functional Maze game where I have learned 2D and 3D UI, waypoint based navigation, procedural animation, interactive object, spatial audio, particle effects, and persistent storage of session data. In this game Players will navigate the maze, collect coins along the way, then find a key that opens a gate to secret treasure.